+++ backend/tests/test_library.c
@@ -847,7 +847,7 @@ static void test_clear(void) {
 
     assert_nonzero(symbol->rows, "ZBarcode_Buffer() rows 0\n");
     assert_nonzero(symbol->width, "ZBarcode_Buffer() width 0\n");
-    assert_zero(symbol->bitmap_byte_length, "ZBarcode_Buffer() bitmap_byte_length %d != 0\n", symbol->bitmap_byte_length);
+    assert_zero(symbol->bitmap_byte_length, "ZBarcode_Buffer() bitmap_byte_length %d != 0\n", (int) symbol->bitmap_byte_length);
     assert_null(symbol->vector, "ZBarcode_Buffer() vector != NULL\n");
 
     ZBarcode_Clear(symbol);
@@ -859,7 +859,7 @@ static void test_clear(void) {
 
     assert_zero(symbol->rows, "ZBarcode_Buffer() rows %d != 0\n", symbol->rows);
     assert_zero(symbol->width, "ZBarcode_Buffer() width %d != 0\n", symbol->width);
-    assert_zero(symbol->bitmap_byte_length, "ZBarcode_Buffer() bitmap_byte_length %d != 0\n", symbol->bitmap_byte_length);
+    assert_zero(symbol->bitmap_byte_length, "ZBarcode_Buffer() bitmap_byte_length %d != 0\n", (int) symbol->bitmap_byte_length);
     assert_null(symbol->vector, "ZBarcode_Buffer() vector != NULL\n");
 
     // Vector
@@ -883,7 +883,7 @@ static void test_clear(void) {
     assert_nonzero(symbol->width, "ZBarcode_Buffer_Vector() width 0\n");
     assert_null(symbol->bitmap, "ZBarcode_Buffer_Vector() bitmap != NULL\n");
     assert_null(symbol->alphamap, "ZBarcode_Buffer_Vector() alphamap != NULL\n");
-    assert_zero(symbol->bitmap_byte_length, "ZBarcode_Buffer_Vector() bitmap_byte_length %d != 0\n", symbol->bitmap_byte_length);
+    assert_zero(symbol->bitmap_byte_length, "ZBarcode_Buffer_Vector() bitmap_byte_length %d != 0\n", (int) symbol->bitmap_byte_length);
 
     ZBarcode_Clear(symbol);
 
@@ -891,7 +891,7 @@ static void test_clear(void) {
 
     assert_zero(symbol->rows, "ZBarcode_Buffer_Vector() rows %d != 0\n", symbol->rows);
     assert_zero(symbol->width, "ZBarcode_Buffer_Vector() width %d != 0\n", symbol->width);
-    assert_zero(symbol->bitmap_byte_length, "ZBarcode_Buffer_Vector() bitmap_byte_length %d != 0\n", symbol->bitmap_byte_length);
+    assert_zero(symbol->bitmap_byte_length, "ZBarcode_Buffer_Vector() bitmap_byte_length %d != 0\n", (int) symbol->bitmap_byte_length);
     assert_null(symbol->bitmap, "ZBarcode_Buffer_Vector() bitmap != NULL\n");
     assert_null(symbol->alphamap, "ZBarcode_Buffer_Vector() alphamap != NULL\n");
     assert_zero(symbol->bitmap_width, "ZBarcode_Buffer_Vector() bitmap_width %d != 0\n", symbol->bitmap_width);
